% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_db.R
\name{call_db}
\alias{call_db}
\title{Call database}
\usage{
call_db(data, prenoms)
}
\arguments{
\item{data}{data form DB}

\item{prenoms}{prenoms to filter}
}
\value{
view of result
}
\description{
Call database
}
