% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the good database}
\usage{
connect_db(
  dbname = "rr2021",
  host = "127.0.0.1",
  port = 5432,
  user = "user",
  password = "password"
)
}
\arguments{
\item{dbname}{name db}

\item{host}{host}

\item{port}{port}

\item{user}{user}

\item{password}{password}
}
\value{
connect
}
\description{
Connect to the good database
}
