% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_evol.R
\name{graph_evol}
\alias{graph_evol}
\title{Graphique sur l'evolution d'un prenoms dans le temps}
\usage{
graph_evol(data, prenoms)
}
\arguments{
\item{data}{deja filter avec le prenom}

\item{prenoms}{prenoms à afficher}
}
\value{
ggplot
}
\description{
Graphique sur l'evolution d'un prenoms dans le temps
}
\examples{
\dontrun{
library(dplyr)
data <- prenoms::prenoms \%>\%
 filter(name == "Vincent" ) \%>\% 
 group_by(year) \%>\% 
 summarise(n = sum(n))

graph_evol(data, "Vincent")}
}
