% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the good database}
\usage{
connect_db(
  dbname = "itdd",
  host = "127.0.0.1",
  port = 5432,
  user = "user",
  password = "password"
)
}
\arguments{
\item{dev}{TRUE or FALSE}
}
\value{
connect
}
\description{
Connect to the good database
}
